requirement.txt
panda 
streamlit
